fixed namespace bug 1 of 2
I believe it is better here to pass in ‘add_objects’ as the first (…) argument to save(), this way R looks up the search path for its definition.  When it’s passed as the ‘list’ argument, I believe R only looks in the scope of slurm_apply().

I’m not a namespace expert, so take this all with a grain of salt.  It’s what I needed to do to get my rslurm-dependent code to work.

Signed-off-by: Robert Corty <rcorty@gmail.com>