fix namespace bug 2 of 2
I believe as a result of my previous change, the capture.output(f) was tagged with a slightly longer and more complex namespace.  This regex will match the simpler, 2-part tag that was there before, as well as the more complex, 3-part tag that seems to be showing up now.

Signed-off-by: Robert Corty <rcorty@gmail.com>